refactor
Simulation.h now has Simulation.cpp
